web/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
dnspython==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
certifi==2023.11.17
APScheduler==3.10.4
jinja2==3.1.2
httpx==0.25.1
pymongo==4.5.0